+++ DonDog-iOS/DonDog-iOS/Presentation/Features/OnBoarding/InitialOnBoardingView.swift
@@ -71,6 +71,68 @@ struct InitialOnBoardingView: View {
     }
 }
 
+struct InitialOnBoardingContent {
+    let imageName: String
+    let title: String
+    let description: String
+    let buttonTitle: String
+}
+
+enum InitialOnBoardingStep: Int, CaseIterable {
+    case first
+    case second
+}
+
+struct InitialOnBoardingContentView: View {
+    let content: InitialOnBoardingContent
+    let onNext: () -> Void
+    let screenWidth = UIScreen.main.bounds.width
+    
+    var body: some View {
+        ZStack(alignment: .topLeading) {
+            Color.ppWhite
+                .ignoresSafeArea()
+            
+            VStack {
+                Image(content.imageName)
+                    .resizable()
+                    .scaledToFill()
+                    .frame(width: UIScreen.main.bounds.width)
+                
+                VStack(spacing: 68) {
+                    VStack(spacing: 10) {
+                        Text(content.title)
+                            .font(.titleBold24)
+                            .foregroundStyle(.ppGray700)
+                        
+                        Text(content.description)
+                            .font(.bodyRegular16)
+                            .foregroundStyle(.ppGray600)
+                    }
+                    .multilineTextAlignment(.center)
+                    
+                    CustomButton(title: content.buttonTitle, action: onNext)
+                        .padding(.horizontal, 20)
+                }
+                .padding(.top, 22)
+            }
+        }
+        // .ignoresSafeArea(.container, edges: .vertical)
+    }
+}
+
+#Preview {
+    InitialOnBoardingContentView(
+        content: InitialOnBoardingContent(
+            imageName: "OnBoardingStep1",
+            title: "사진 두 장이면 충분해요",
+            description: "사랑하는 가족과의 대화, 부담으로 느껴질 때가 있죠.\n사진만으로 가볍게 안부를 전해요.",
+            buttonTitle: "다음으로"
+        ),
+        onNext: {}
+    )
+}
+
 #Preview {
     InitialOnBoardingView()
 }
